requirements.txt
pytest==7.2.0
PyGithub==1.57